+++ model/schema_insert_test.go
@@ -8,3 +8,10 @@ func TestInsert(t *testing.T) {
 	// TODO: Schema.InsertTx
 	// TODO: Schema.InsertTxAt
 }
+
+func TestInsertRows(t *testing.T) {
+	// TODO: Schema.InsertRows
+	// TODO: Schema.InsertRowsAt
+	// TODO: Schema.InsertTxRows
+	// TODO: Schema.InsertTxRowsAt
+}